Handle the case of a brand new Git install failing to run git config --global
When a fresh install of git happens on a system, the git global config
file does not exist yet. This results in git config --global -l failing.
In this case, simply return an empty map instead of null.

Additionally, if an empty map is detected, set a WTMP-specific parameter
to "false" (the ignoreSchannel property), and check the config again. If
it hasn't been added to the config, print a warning that
autoconfiguration may not function properly, since updating the config
does not seem to work as expected.

Additionally, use `git --version` instead of `git --help` to determine
if git is installed, and print that version out at most once per day.